\name{aspep}
\alias{aspep2007_census}
\alias{aspep2009_sample}
\alias{aspep2010_sample}
\alias{aspep2011_sample}
\alias{aspep2012_sample}
\alias{aspep2013_census}
\docType{data}
\title{aspep}
\description{

}
\usage{data(aspep2007_census)}
\format{
 These tables contain data from the Annual Survey of Public employment and Payroll.
  \describe{
    \name{aspepXXXX_}
\item{ ID }{ ID }
\item{ state }{ state }
    \item{itemcode}{(Item code. "000": summary, 
001= Airport
007= Space Research and Technology (Federal),
005= Correction,
006= National Defense and International Relations(Federal),
044= Streets and Highways,
050= Housing and Community Development (Local),
052= Local Libraries,
059= Natural Resources,
061= Parks and Recreation,
012= Elementary and Secondary - Instruction,
112= Elementary and Secondary - Other Total,
014= Postal Service (Federal),
016= Higher Education - Other,
018= Higher Education - Instructional,
062= Police Protection - Officers,
162= Police-Other,
129= Welfare,
080= Sewerage,
071= Other Education (State) 081 Solid 072, Social Insurance Administration (State),
073= Financial Administration,
074= Firefighters,
124= Fire - Other,
081= Waste Management,
087= Water Transport and Terminals,
089= Other and Unallocable,
090= Liquor Stores (State),
091= Water Supply,
075= Judicial and Legal,
079= Other Government Administration,
032= Health,
092= Electric Power,
093= Gas Supply,
094= Transit)
}
\item{ ftemp09 }{ ftemp09 }
\item{ ftpay09 }{ ftpay09 }
\item{ ptemp09 }{ ptemp09 }
\item{ ptpay09 }{ ptpay09 }
\item{ pthours09 }{ pthours09 }
\item{ pefCode }{ pefCode }
\item{ finalWeight }{ finalWeight }
\item{ ftemp07 }{ ftemp07 }
\item{ ftpay07 }{ ftpay07 }
\item{ ptemp07 }{ ptemp07 }
\item{ ptpay07 }{ ptpay07 }
\item{ pthours07 }{ pthours07 }
\item{ inSample }{ inSample }
\item{ littleN }{ littleN }
\item{ bigN }{ bigN }
\item{ substratum }{ substratum }
}
}
\source{\url{https://www.census.gov//govs/apes/}}
\keyword{datasets}
  