\name{employment_id}
\alias{employment_id}
\docType{data}
\title{employment_id}
\description{
  
}
\usage{data(employment_id)}
\format{
  \describe{
    \name{employment_id}
    \item{id}{Id}
    \item{state}{State}
    \item{samplingType}{Type with levels "2" and "3" collapsed: level "3" corresponds to townships and Municipalities}
    \item{insample12}{inSample12}
    \item{totalpay07}{totalpay07}
    \item{pik}{pik}
  }
}
\source{

}
\examples{
}
\keyword{datasets}
