\name{employment_strata}
\alias{employment_strata}
\docType{data}
\title{employment_strata}
\description{
  
}
\usage{data(employment_strata)}
\format{
  \describe{
    \name{employment_strata}
    \item{state}{State}
    \item{samplingType}{Type with levels "2" and "3" collapsed: level "3" corresponds to townships and Municipalities}
    \item{Nh}{Nh}
    \item{sumsizevar}{sumsizevar}
  }
}
\source{

}
\examples{
}
\keyword{datasets}
